Bug Fixed and polished for Prototype
Fixed the shooting and health bar position. Added game won and game over scenes with a transition to game over when the ally dies. Friction is still broken but there is a plan to fix it. Changed the colliders to be circles so they bounce in more interesting ways. Lastly I need to unify the positioning of everything to be Vector3.